+++ platforms/jvm/toolchains-jvm/src/test/groovy/org/gradle/jvm/toolchain/install/internal/SecureFileDownloaderTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.jvm.toolchain.install
+package org.gradle.jvm.toolchain.install.internal
 
 import org.gradle.api.Action
 import org.gradle.api.BuildCancelledException
